+++ src/lib/tasks/init.ts
@@ -1,0 +1,25 @@
+import { InitResult } from '../../../typings';
+import { StringTask } from './task';
+import { parseInit } from '../responses/InitSummary';
+
+const bareCommand = '--bare';
+
+function hasBareCommand(command: string[]) {
+   return command.includes(bareCommand);
+}
+
+export function initTask(bare = false, customArgs: string[]): StringTask<InitResult> {
+   const commands = ['init', ...customArgs];
+   if (bare && !hasBareCommand(commands)) {
+      commands.splice(1, 0, bareCommand);
+   }
+
+   return {
+      commands,
+      concatStdErr: false,
+      format: 'utf-8',
+      parser(text: string): InitResult {
+         return parseInit(commands.includes('--bare'), text);
+      }
+   }
+}